+++ src/packages/tyhUi/packages/list/src/prop.ts
@@ -1,0 +1,12 @@
+export const prop = {
+  header: String,
+  content: {
+    type: Array,
+    default: () => []
+  },
+  keys: String,
+  footer: String,
+  zebra: Boolean,
+  num: Boolean,
+  hoverShow: Boolean
+}